Fixes a variety of issues found in the codebase while manually running through the flow.
1. Changed way the business name was being accessed from the DataRow for HMRC CSV data.

Uses row['Business Name'] instead of row.BUSINESS_NAME; matching the format of it within the actual object.

Fixes issues with HMRC upload process failing.

2. Snapshot manager was not correctly identifying old snapshot tables to delete past the limit of 5. This was because the name passed into the query being wrong.

I have now returned the actual basename of the table, and have constructed a proper snapshot table name from this.

Highlights how the testing for the Snapshot module may not accurately test for the deletion of old tables, past the default limit of 5.

Though, due to time constraints I don't think it's possible for me to rework the tests. A point for my reflective report I feel.

In short, oldest snapshot past the limit are now deleted.

3. One of the Scraping components methods was making an additional call to update using the Gambling Commission CSVs.

This was unintended, as it should only happen once on the initial load, not twice.

Now fixed so it will perform the update once.

4. I incorporated an anonymous async arrow function that will auto-execute on application load.

Making sure to use await, so the creation of the schema is properly awaited before moving onto the next set-up steps.

+++ src/components/TableSnapshot/SnapshotManager.ts
@@ -22,12 +22,18 @@ export default class SnapshotManager {
 
 	private async createSnapshotTable(regulator: Regulator): Promise<string> {
 		const baseTable = this.getBaseTableName(regulator);
+
+		// Shouldn't ever throw, just here to remove the eslint issue.
+		if (!baseTable) {
+			throw new Error('Undefined base table. Invalid regulator provided.');
+		}
+
 		const snapshotName = `${baseTable}_snapshot_${Date.now()}`;
 
 		await this.client.query(`CREATE TABLE ${this.schema}.${snapshotName} AS TABLE ${this.schema}.${baseTable}`);
 		console.log(`Snapshot created: ${snapshotName}`);
 
-		return snapshotName;
+		return baseTable;
 	}
 
 	private getBaseTableName(regulator: string) {
@@ -59,18 +65,21 @@ export default class SnapshotManager {
            to oldest.
 
            Finally, we use OFFSET to ignore the top 5 recent results, and retain only the excess tables (snapshots) that should be deleted. */
+		const snapshotTableName = `${tableName}_snapshot_%`;
 		const excessSnapshots = await this.client.query(`
             SELECT table_name
             FROM information_schema.tables
-            WHERE table_schema = $1 AND table_name LIKE $2
-            ORDER BY table_name DESC
-            OFFSET ${maxSnapshots}`,
-		[this.schema, `${tableName}_snapshot_%`]);
+			WHERE table_schema = $1 AND table_name ILIKE $2
+			ORDER BY table_name DESC
+			OFFSET $3;`,
+		[this.schema, snapshotTableName, maxSnapshots]);
 
+		if (excessSnapshots.rows) {
 		// Delete excess tables.
-		for await (const snapshot of excessSnapshots.rows) {
-			await this.client.query(`DROP TABLE ${this.schema}.${snapshot.table_name}`);
-			console.log(`Deleted old snapshot: ${snapshot.table_name}`);
+			for await (const snapshot of excessSnapshots.rows) {
+				await this.client.query(`DROP TABLE ${this.schema}.${snapshot.table_name}`);
+				console.log(`Deleted old snapshot: ${snapshot.table_name}`);
+			}
 		}
 	}
 }
